chore: add exports and types
