+++ projects/ion/src/lib/core/types/index.ts
@@ -19,6 +19,7 @@ export * from './message';
 export * from './notification';
 export * from './pagination';
 export * from './popconfirm';
+export * from './sidebar';
 export * from './simple-menu';
 export * from './smart-table';
 export * from './status';
@@ -27,4 +28,3 @@ export * from './tab-group';
 export * from './table';
 export * from './tag';
 export * from './tooltip';
-export * from './indicator';
